Allow using different namespace for BMH
This would allow for using the `openstack` or any other namespace
to look for BaremetalHosts. Though we should create BMHs in the
default `openstack` namespace and not in `openshift-machine-api`,
keeping the default would provide backward compatibilty with OSPdO
and if OpenShift Assisted Installer (AI) and BMHs created in
`openshift-machine-api` namespace.