Make some CRD fields optional
These are used to configure ctlplane network. We possibly don't
need to keep all of them as required. Also we use
OpenStackBaremetalSetSpec directly in dataplane and these fields
would be popluted with IPAM and won't be part of the datplane CRD.
We could use a slimmed down version of OpenStackBaremetalSetSpec
inline in dataplane CRD though.