Enable baremetal deployment with virtualmedia
When the bootMode in provisioning resource is either `Unmanaged`
or Disabled, we would notlook for the provisioning interface and
rather boot ramdisk using virtualmedia.Also, use the hostIP to
build the local image url. The only caveat here is that BM network
where the os image is available locally should be routable from the OSP
controlplane network, if we use a different network for OSP
controlplane rather than the BM network.