Change initialDelay of startupProbe to the failureThreshold
It seems that 30 seconds which were set before was not enough
to generate SSL certificate and start osp-httpd container in the POD.
Problem was visible for example when this was deployed locally on dev
machine.
To avoid that problem, and don't need to wait for livenessProbe to start
for too long time, this patch removes initialDelay parameter of the
startupProbe and instead sets:

    failureThreshold: 12
    periodSeconds: 10

That way startupProbe can fail 12 times before container will be
restarted. But checks will be performed every 10 seconds and first
successful check will stop startupProbe and start livenessProbe.

As livenessProbe will be always started after successful check of the
startupProbe we don't need initialDelay there at all so it's removed
also.