Remove PreprovisioningNetworkData
We had initially introduced preprovisioningNetworkDataName
field in InstanceSpec for users to have flexibility of
setting it in OpenStackDataPlaneNodeSet CR for the individual
nodes. We have had the implementation to fallback to using
preprovisioningNetworkDataName for networkData if the later
is not provided and the former is set.

However, we've noticed that this does not work when using
coreos IPA ramdisk image as ProvisioningImage Builder expects
the preprovisioningNetworkDataName to be in nmstate format,
however for us during provisioning cloud-init expects it to be in
openstack network_data.json format.

Therefore we would not replicate the metal3 behavior of falling back
to preprovisioningNetworkDataName if networkData is not provided.
We instead will keep both network data for preprovisionig and
provisoning separate. If networkData is not provided per BMH node,
we would use the default networkData we generate using IPAM.

This also removes networkData/userData from the BaremetalSet spec
as there is no way user can provide this data for a set of BMHs.

jira: https://issues.redhat.com/browse/OSPRH-10442
Signed-off-by: rabi <ramishra@redhat.com>