Reconcile for BMH delete events
As one can delete BMHs with consumerRef outside the general compute
scale-down method, so it's better we should reconcile when for instance
faulty BMHs are deleted and replace the compute with a new BMH (if available).

Users just need to recreate a deployment to deploy on those newly
created nodes.

Signed-off-by: rabi <ramishra@redhat.com>