Use instance name and namespace for existingPorts map key
It's possible to have provisionServers of the same name in
different namespaces. This change ensures that we use the instance
namespace while formatting the key for the existingPorts map. This
will ensure we avoid duplicates if the same names are used in multiple
namespaces.

Jira: https://issues.redhat.com/browse/OSPRH-15769
Signed-off-by: Brendan Shephard <bshephar@redhat.com>

Use instanceKey

Signed-off-by: Brendan Shephard <bshephar@redhat.com>